Fix priority label in Board page
